@@ -30,7 +30,7 @@ func New(p ppi.Plugin) *cobra.Command {
 		Use:   Name + " <name> <spec>",
 		Short: "execute a value merge",
 		Long: `
-This command executes a a value merge. The values are taken from *stdin* as JSON
+This command executes a value merge. The values are taken from *stdin* as JSON
 string. It has the following fields: 
 
 - **<code>local</code>** *any*
@@ -39,7 +39,7 @@ string. It has the following fields:
 
 - **<code>inbound</code>** *any*
 
-  The value to merge into. THis value is based on the original inbound value.
+  The value to merge into. This value is based on the original inbound value.
 
 This action has to provide an execution result as JSON string on *stdout*. It has the 
 following fields: 
